update_metadata now respects tc.py's DB_NAME
Changed update_metadata so that it looks up the DB_NAME from tc.py. This will
make it a lot easier to refactor that database name, if necessary. Also removed
timeclock.db from the list of tracked files and added it to .gitignore, because
honestly we don't need to put data in version control.

* Pointed update_metadata at tc's DB_NAME
* Removed timeclock.db from VCS
  - run `$ python update_metadata.py` to restore a working copy